# weather_service/internal/response/response.go
package response

import (
	"github.com/gin-gonic/gin"
	"github.com/roxyash/go_ahead/pkg/zaplogger"
)

type ForecastResponse struct {
	Temp string `json:"temp"`
	City string `json:"city"`
}

type ErrorResponse struct {
	Message string `json:"message"`
}

type AboutResponse struct {
	Message string `json:"message"`
}

type StatusResponse struct {
	Status string `json:"status"`
}

type Response struct {
	logger zaplogger.Logger
}

func NewResponse(logger zaplogger.Logger) *Response {
	return &Response{
		logger: logger,
	}
}

func (r *Response) NewErrorResponse(c *gin.Context, statusCode int, message string) {
	r.logger.Errorf(message)
	c.AbortWithStatusJSON(statusCode, ErrorResponse{Message: message})
}